requirements.txt
# Core dependencies
redis>=5.0.0
pydantic>=2.0.0
asyncio-redis>=0.16.1
python-dotenv>=1.0.0

# DSPy for LLM abstraction
dspy-ai>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0

# Logging and monitoring
structlog>=23.0.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0

# Type checking
mypy>=1.0.0
types-redis>=4.0.0

# Monitoring and configuration
PyYAML>=6.0.0
psutil>=5.9.0

# Load testing
aiohttp>=3.8.0  # For HTTP load testing scenarios